separating out NumberInputField from NumberComponent (to be used in MethodComponent)
